Documentation of kernelBuilder continued.
  The completion of the SW documentation unveiled the inconsistent use of
the terms starting and creating of contexts. Resolving this wording issue
required renaming of the functions for creation of the context save
descriptor objects. Previously, the function names promised to create a
context, but at least for single-shot contexts they effectively create a
descriptor only. For single-shot contexts, the descriptor is typically
created once, while the context is created repeatedly and frequently -
always from one and the same descriptor